requirements-dev.txt
Django==2.1.2
django-ajax-selects==1.7.1
django-filter==2.0.0
django-reset-migrations==0.3.1
django-rest-auth==0.9.3
django-seed==0.1.9
djangorestframework==3.8.2
Faker==0.9.1
future==0.16.0
gunicorn==19.9.0
httplib2==0.9.2
idna==2.7
Jinja2==2.10
jinja2-time==0.2.0
keyring==10.6.0
keyrings.alt==3.0
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
Markdown==3.0.1
MarkupSafe==1.0
mysqlclient==1.3.13
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
pipenv==2018.7.1
poyo==0.4.1
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0

python-dateutil==2.7.3
python-debian==0.1.32
pytz==2018.5
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.19.1
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0

systemd-python==234
text-unidecode==1.2
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
wadllib==1.3.2

zope.interface==4.3.2
dj-static==0.0.6
dj-database-url==0.5.0
python-decouple==3.1
psycopg2==2.7.5
